edited - overall organized, created re-usable components
