+++ src/components/article/styles/articleDetail.style.scss
@@ -1,4 +1,4 @@
-.modal {
+.article__detail {
   position: fixed;
   width: 100%;
   height: 100%;
@@ -10,13 +10,12 @@
   padding: 0.75em;
   z-index: 999;
 
-  &__content {
+  &-content {
     display: flex;
     max-width: 40em;
     max-height: fit-content;
     width: 100%;
     height: auto;
-    padding: 1em;
     overflow-y: auto;
     height: 100%;
     background-color: var(--clr-bg-lgt);
@@ -42,12 +41,12 @@
     //----------------------------------//
   }
 
-  &__close {
+  &-close {
     position: absolute;
     right: -0.1rem;
     top: -0.1rem;
     width: 4em;
-    background-color: #454954;
+    background-color: #1f232d;
     padding: 0.25em;
     cursor: pointer;
     z-index: 9;
@@ -55,7 +54,7 @@
     border-bottom-left-radius: 8px;
 
     &:hover {
-      background-color: #191c24;
+      background-color: #010101;
     }
 
     svg {
@@ -76,18 +75,15 @@
     color: var(--clr-font-drk);
     font-weight: bold;
   }
-}
 
-/* ------------ Description starts here ------------ */
-.modal__desc {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-
+  &-desc {
+    display: flex;
+    flex-direction: column;
+    gap: 1em;
+  }
   .image-container {
     width: 100%;
     aspect-ratio: 3/2;
-    border-radius: 4px;
     overflow: hidden;
   }
   &-info {
@@ -102,11 +98,13 @@
     h3 {
       color: var(--clr-font-drk);
       font-weight: bold;
+      padding: 0 0.75em;
     }
     p {
       color: var(--clr-font);
       display: flex;
       flex-direction: column;
+      padding: 0 0.75em;
     }
   }
   &-source {
@@ -118,6 +116,8 @@
   &-link {
     display: flex;
     flex-direction: column;
+    padding: 0 0.75em;
+
     a {
       color: var(--clr-font-link);
       margin-right: auto;
@@ -129,6 +129,7 @@
     display: flex;
     justify-content: space-between;
     color: var(--clr-font);
+    padding: 0 0.75em;
   }
   &-stats {
     display: flex;
