+++ src/components/article/styles/articleGrid.style.css
@@ -1,0 +1,18 @@
+.article {
+  display: flex;
+  flex-direction: column;
+}
+.article__header {
+  height: 5em;
+  display: flex;
+  align-items: center;
+}
+.article h1 {
+  font-weight: bolder;
+  font-size: 1.6em;
+}
+.article__content {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: 2em;
+}/*# sourceMappingURL=articleGrid.style.css.map */
