+++ src/pages/HomePage/home.style.css
@@ -111,115 +111,8 @@
   color: var(--clr-font-drk);
 }
 
-.home__landing {
-  display: flex;
-  flex-direction: column;
-}
-.home__landing h1 {
-  font-weight: bolder;
-  font-size: 1.6em;
-}
-.home__landing-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 2em;
-}
-.home__landing-text:hover h2,
-.home__landing .image-container:hover ~ .home__landing-content .home__landing-text h2 {
-  text-decoration: underline;
-}
-.home__landing h2 {
-  color: var(--clr-font);
-  font-weight: bold;
-}
-.home__landing p {
-  color: var(--clr-font-md);
-}
-.home__landing .image-container {
-  aspect-ratio: 3/2;
-  width: 100%;
-  border-radius: 1em;
-  overflow: hidden;
-  flex-shrink: 0;
-  cursor: pointer;
-}
-.home__landing .no-image {
-  width: 100%;
-  aspect-ratio: 3/2;
-  background-color: rgb(57, 56, 130);
-}
-.home__landing-card {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-}
-.home__landing-content {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  height: 100%;
-}
-.home__landing-text {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-  cursor: pointer;
-}
-.home__landing-details {
-  display: flex;
-  gap: 1em;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: auto;
-}
-.home__landing-date {
-  display: flex;
-  align-items: center;
-  color: var(--clr-font);
-}
-.home__landing-stats {
-  display: flex;
-  gap: 1em;
-}
-.home__landing-stats button {
-  display: flex;
-  align-items: center;
-  gap: 0.5em;
-  font-weight: bold;
-  color: var(--clr-font);
-}
-.home__landing-stats button:hover svg {
-  color: var(--clr-font);
-}
-.home__landing-stats svg {
-  fill: none;
-  stroke-width: 1px;
-  color: var(--clr-font-lgt);
-}
-.home__landing-empty {
-  width: 100%;
-  margin: auto;
-  min-height: 50vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 2em;
-}
-.home__landing-empty p {
-  font-size: 1.1em;
-}
-.home__landing-empty .image-container {
-  width: 10rem;
-  height: 10rem;
-}
-
 @media (max-width: 920px) {
   .home__aside {
     display: none;
-  }
-}
-@media (max-width: 470px) {
-  .home__landing-empty {
-    flex-direction: column;
   }
 }/*# sourceMappingURL=home.style.css.map */
