+++ src/pages/HomePage/home.style.scss
@@ -130,131 +130,8 @@
   }
 }
 
-//-------- LANDING STYLING STARTS HERE --------//
-.home__landing {
-  display: flex;
-  flex-direction: column;
-
-  h1 {
-    font-weight: bolder;
-    font-size: 1.6em;
-  }
-  &-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 2em;
-  }
-
-  &-text:hover h2,
-  .image-container:hover ~ .home__landing-content .home__landing-text h2 {
-    text-decoration: underline;
-  }
-
-  h2 {
-    color: var(--clr-font);
-    font-weight: bold;
-  }
-  p {
-    color: var(--clr-font-md);
-  }
-
-  .image-container {
-    aspect-ratio: 3/2;
-    width: 100%;
-    border-radius: 1em;
-    overflow: hidden;
-    flex-shrink: 0;
-    cursor: pointer;
-  }
-
-  .no-image {
-    width: 100%;
-    aspect-ratio: 3/2;
-    background-color: rgb(57, 56, 130);
-  }
-
-  &-card {
-    display: flex;
-    flex-direction: column;
-    gap: 1em;
-  }
-
-  &-content {
-    display: flex;
-    flex-direction: column;
-    gap: 1em;
-    height: 100%;
-  }
-
-  &-text {
-    display: flex;
-    flex-direction: column;
-    gap: 1em;
-    cursor: pointer;
-  }
-
-  &-details {
-    display: flex;
-    gap: 1em;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: auto;
-  }
-
-  &-date {
-    display: flex;
-    align-items: center;
-    color: var(--clr-font);
-  }
-
-  &-stats {
-    display: flex;
-    gap: 1em;
-    button {
-      display: flex;
-      align-items: center;
-      gap: 0.5em;
-      font-weight: bold;
-      color: var(--clr-font);
-      &:hover svg {
-        color: var(--clr-font);
-      }
-    }
-    svg {
-      fill: none;
-      stroke-width: 1px;
-      color: var(--clr-font-lgt);
-    }
-  }
-
-  &-empty {
-    width: 100%;
-    margin: auto;
-    min-height: 50vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 2em;
-
-    p {
-      font-size: 1.1em;
-    }
-
-    .image-container {
-      width: 10rem;
-      height: 10rem;
-    }
-  }
-}
-
 @media (max-width: 920px) {
   .home__aside {
     display: none;
-  }
-}
-
-@media (max-width: 470px) {
-  .home__landing-empty {
-    flex-direction: column;
   }
 }
